// Angular/src/app/models/porudzbina.ts
export interface Porudzbina{

    id_porudzbina?: string; 

    datum_kreiranja?: Date;

    ukupna_cena?: number; 

    status_porudzbine?: string; 

    status_placanja?: string; 

    dobavljac?: string; 

    broj_posiljke?: string; 

    rok_isporuke?: Date; 

    id_korisnik?: string; 

    id_dostava?: string;

}